+++ BackEnd/app/Http/Controllers/Admin/SupplierController.php
@@ -21,6 +21,7 @@ class SupplierController extends Controller
 
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Supplier::class);
         $suppliers = $this->supplierService->all($request);
         return view('admin.suppliers.index',compact('suppliers'));
     }
@@ -30,6 +31,7 @@ class SupplierController extends Controller
 
     public function create()
     {
+        $this->authorize('create', Supplier::class);
         return view('admin.suppliers.add');
     }
 
@@ -49,6 +51,7 @@ class SupplierController extends Controller
 
     public function edit($id)
     {
+        $this->authorize('update', Supplier::class);
         $item = $this->supplierService->find($id);
         return view('admin.suppliers.edit',compact('item'));
     }
@@ -68,6 +71,7 @@ class SupplierController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('delete', Supplier::class);
         try {
             $category = $this->supplierService->delete( $id);
             Session::flash('success', config('define.recycle.succes'));
@@ -84,7 +88,9 @@ class SupplierController extends Controller
         return view('admin.suppliers.recycle',compact('suppliers'));
     }
 
-    public function restore($id){
+    public function restore($id)
+    {
+        $this->authorize('delete', Supplier::class);
         try {
             $this->supplierService->restore($id);
             Session::flash('success', config('define.restore.succes'));
@@ -96,7 +102,9 @@ class SupplierController extends Controller
         }
     }
 
-    public function force_destroy($id){
+    public function force_destroy($id)
+    {
+        $this->authorize('forceDelete', Supplier::class);
         try {
             $category = $this->supplierService->force_destroy( $id);
             Session::flash('success', config('define.delete.succes'));
